Changed (int) to Int64 in domain title_ID. Added knownCharacters in CharacterController (works but, does not add bool value to Known_For

+++ Rawdata_Porfolio_2/Entity Framework/Domain/bookmarks_title.cs
@@ -10,7 +10,7 @@ namespace Rawdata_Porfolio_2.Pages.Entity_Framework.Domain
         [Key]
         public int User_Id { get; set; }
         public User User { get; set; }
-        public int Title_Id { get; set; }
+        public Int64 Title_Id { get; set; }
         public Title Title { get; set; }
         public string Note { get; set; }
         public DateTime Timestamp { get; set; }
