+++ Rawdata_Porfolio_2/Entity Framework/Domain/title_genres.cs
@@ -9,7 +9,7 @@ namespace Rawdata_Porfolio_2.Pages.Entity_Framework.Domain
     public class Title_Genre
     {
         [Key]
-        public int Title_Id { get; set; }
+        public Int64 Title_Id { get; set; }
         public Title Title { get; set; }
         public string Genre { get; set; }
